Variable Population Cap Support Added
It is now possible to set a specific population cap in LevelData.xml.
Add:
<StandardPop Value="[desired population]" />
to a level description. Defaults to 15 if nothing is specified.
The absolute cap on a level is that value + 3, so if the player presses
HitMe, she has to make at least one match before HitMe will be activated
again (if there are still more cards in the "deck" off screen).